requirements.txt
asgiref==3.6.0
Django==4.1.7
djangorestframework==3.14.0
pytz==2022.7.1
sqlparse==0.4.3
djangorestframework_simplejwt==5.2.2
pyjwt==2.6.0
